src/app/feature/products/page/product/product.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { By } from '@angular/platform-browser';
import { ProductComponent } from './product.component';
import { MockStore, provideMockStore } from '@ngrx/store/testing';
import { of } from 'rxjs';
import { loadProducts } from '../../store/products.actions';
import { ProductCardComponent } from '../../components/product-card/product-card.component';

describe('ProductComponent', () => {
  let component: ProductComponent;
  let fixture: ComponentFixture<ProductComponent>;
  let store: MockStore;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ProductComponent],
      providers: [provideMockStore()]
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(ProductComponent);
    component = fixture.componentInstance;
    store = TestBed.inject(MockStore);

    spyOn(store, 'dispatch');
    component.ngOnInit();
    expect(store.dispatch).toHaveBeenCalledWith(loadProducts());

    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should fetch the products on initialization', () => {
    expect(store.dispatch).toHaveBeenCalledWith(loadProducts());
  });

  it('should render the app-product-card component with the products', () => {
    const productCardComponent = fixture.debugElement.query(By.css('app-product-card'));
    expect(productCardComponent).toBeTruthy();
  });
});
